Resolving PR comments
